requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
pytest==7.2.1
python-dotenv==0.21.1
spotipy==2.22.0
